Better ignore PCR duplicates option handling. Still need to add option to config file.
Now flag is called IGNOREPCRDUP. pcr_dup C program now takes in a 5th argument
which it interprets as this flag. If it is any value considered to be "TRUE"
the conditional branch which would lead to the removal of a sequence considered
a PCR duplicate is never taken. pcr_dup.pl also takes an extra argument
containing the value of this flag so it can be passed to pcr_dup.exe