(unstable, testing) Reference TR db will only be attached when necessary, begin removal of MySQL-specific options and code
- Reference DB will now only be attached when requested using "userefdb"
key in hashref argument to get_dbh.
- Fixed all instances where get_dbh was improperly given the run db
location instead of the (optional) hashref of parameters.
- Added ability to connect to run db as a read-only database. Useful
if we want to make sure the db is not updated or locked when attaching
a shared reference db.
- Now exporting get_ref_dbh which can be used by callers to get reference
db only. Useful for scripts which only need to interact with the ref db.
- Consolidated all functions which modify the reference db into one
initializer function. Each function still detects whether or not an
update is needed, and can be forced to update with the "redo" parameter.
- get_ref_dbh needs to be given all parameters needed to initialize
or connect to the reference db so that utility scripts, which do not
need to initialize a hash with run parameters, can use it as well.
- Save the reference DB (REFERENCE_DB) location into the run parameter
hash so that, at least within a run, the location is easily accessible.
- Removed many unneeded parameters which we used only for MySQL
(vutil.pm, vs.cnf)
- Removed some code which was MySQL-specific.
- updaterefs.pl now takes fewer arguments, since most are available
from the run configuration hash, to which it has access.
- updaterefs.pl has a new function, which bundles all the code used
to update the run-specific reference table (used for tallying results)
- Hopefully friendlier/less misleading message from TRF2PROCLU when
no TRs were processed by TRF.