Fixed bug when splitting work for trf which resulted in misleading error messages and extraneous files
- Fixed a bug in ProcInputReads.pm which was caused by a postincrement
error. Condition should have checked variable AFTER thevariable was
incremented. However, a postincrement was used instead. The result
was that a TRF pipe was terminated after the very first read was
processed, leading to most first splits being empty (unless there
happened to be a TR in the first read). This led to many error
messages from trf2proclu indicating no TRs found in input.
- Changed all strings used for int array indexing in trf2proclu source
files to be unsigned char *. If they were ever used in a context where
the signness did not matter and it was safe, they were cast to char *.
- Using safer min/max macros in trf2proclu source. May extend use in
other C programs.
- Gracefully handle using older testing config files which used old
FASTA_DIR option.