Changed file name requirements; better way to ensure unique read names; default option change
- Now looks for format name anywhere in file name
- Add index field in fasta header if reading a FASTA/FASTQ file where we
cannot determine the member of the pair for paired end reads. Actually
does this for unpaired too, but doesn't cost much extra. This helps
distinguish reads when the headers are identical for both members of
a pair.
- KEEPPCRDUPS option is now on by default. This means that unless changed
by the user, step 15 is not run. Note that you will need to overwrite
your global vs.cnf, or edit it to match this change, if you want this
change applied to your config since the installer does not automatically
overwrite the global config.