simpler run init, sane defaults, cmake cleanup/dep fixes, seqtk rm'd
- VNTRseek is now much easier to initialize. Fewer CLI options are
mandatory, and some hopefully sane defaults are now used to save time
and effort.
- Indistinguishable production script also now updated to work with
all the latest changes.
- CMake files have been modernized slightly, and new dependencies are
tracked by CMake.
- seqtk code is no longer included. Instead, CMake will download a
specific version of the code (currently, the latest tagged release)
so that we don't have to bundle it. This does add git and an internet
connection as install-time dependencies (but everyone should have
those, I hope).
- Note: Some minor differences in VCF files might be noticed. This
is limited to the allele representative sequence which is printed to
the VCF possibly being different from earlier versions. Older versions
potentially had a harmless bug where a read other than the
representative was selected for the VCF file, but the selection was
still correct because the read still called the correct allele. Still
some verification with larger runs needed to rule out any actual bugs.