Fixed bug in redund where sqlite db was not created when preparing ref set
redund.c
- Moved string allocation to before single outfile branching point. Now
that the file name variables are dynamically allocated, sprintf does not
work to set them so both branches expect initialized buffers to write to.
- Fixed another bug where an error would occur when reading the DEBUG
env var if it is not set.
- Free filename vars

vntrseek.pl
- Removed unused chdirs
- Moved find command invocation directly inside system call

vutil.pm
- Surround call to run_redund with a try block. Drops the database table
so that this load_refprofiles_db can be run again in case of an error.
- In run_redund, add some code to help figure out any error with
retrieving SQL code for table creation. (Might be removed)